Implementing forgetting Map with test cases
A forgetting Map has the capacity to take in n associations. Once full, to add a new entry to the Map, the least used entry is removed. The least used entry is decided by the number of times the entry is accessed. If multiple entries are accessed for the same number of times, then as this implementation is based on TreeMap which is always sorted based on keys (ascending order), the first key which occurs in the ascending order is removed.